reset fix

+++ reset.sh
@@ -4,23 +4,24 @@ SCRIPT=`realpath "$0"`
 SCRIPT_DIR=`dirname "$SCRIPT"`
 PROJECT_ROOT=$SCRIPT_DIR
 
+echo "Project Root: $PROJECT_ROOT"
 echo "This will reset everything that isn't committed..."
 
-[[ "$(read -e -p 'Continue? [y/N]> '; echo $REPLY)" == [Yy]* ]]
+if [[ "$(read -e -p 'Continue? [y/N]> '; echo $REPLY)" == [Yy]* ]]; then
+    cd $PROJECT_ROOT
 
-cd $PROJECT_ROOT
+    echo "Cleaning rust build data..."
+    cargo clean
 
-echo "Cleaning rust build data..."
-cargo clean
+    echo "Cleaning android build data..."
+    cd $PROJECT_ROOT/android && $PROJECT_ROOT/android/gradlew clean
 
-echo "Cleaning android build data..."
-cd $PROJECT_ROOT/android && $PROJECT_ROOT/gradlew clean
+    echo "Unstaging all files..."
+    git reset
 
-echo "Unstaging all files..."
-git rm -rf --cached $PROJECT_ROOT
+    echo "Cleaning .gitignored files..."
+    git clean -dfX
 
-echo "Cleaning .gitignored files..."
-git clean -dfX
-
-echo "Resetting repo to HEAD..."
-git reset --hard
+    echo "Resetting repo to HEAD..."
+    git reset --hard
+fi
